+++ fabricgov/__init__.py
@@ -1,0 +1,3 @@
+from fabricgov.api import FabricGov
+
+__all__ = ["FabricGov"]